user basic controller created

+++ src/main/java/com/notahmed/catsfinder/repository/UserRepository.java
@@ -3,6 +3,6 @@ package com.notahmed.catsfinder.repository;
 import com.notahmed.catsfinder.models.User;
 import org.springframework.data.repository.ListCrudRepository;
 
-public interface UserRepository extends ListCrudRepository<User, Integer> {
+public interface UserRepository extends ListCrudRepository<User, Long> {
 
 }
